+++ rust-runtime/Cargo.toml
@@ -10,6 +10,7 @@ members = [
     "aws-smithy-eventstream",
     "aws-smithy-http",
     "aws-smithy-http-auth",
+    "aws-smithy-http-client",
     "aws-smithy-http-server",
     "aws-smithy-http-server-python",
     "aws-smithy-http-tower",
