+++ rust-runtime/aws-smithy-http-client/src/hyper_014.rs
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-use crate::client::http::connection_poisoning::CaptureSmithyConnection;
-use crate::client::http::hyper_014::timeout_middleware::HttpTimeoutError;
+use crate::hyper_014::timeout_middleware::HttpTimeoutError;
 use aws_smithy_async::future::timeout::TimedOutError;
 use aws_smithy_async::rt::sleep::{default_async_sleep, AsyncSleep, SharedAsyncSleep};
 use aws_smithy_runtime_api::box_error::BoxError;
+use aws_smithy_runtime_api::client::connection::CaptureSmithyConnection;
 use aws_smithy_runtime_api::client::connection::ConnectionMetadata;
 use aws_smithy_runtime_api::client::connector_metadata::ConnectorMetadata;
 use aws_smithy_runtime_api::client::http::{
@@ -820,7 +820,7 @@ mod timeout_middleware {
 
     #[cfg(test)]
     mod test {
-        use crate::client::http::hyper_014::HyperConnector;
+        use crate::client::HyperConnector;
         use aws_smithy_async::assert_elapsed;
         use aws_smithy_async::future::never::Never;
         use aws_smithy_async::rt::sleep::{SharedAsyncSleep, TokioSleep};
