+++ rust-runtime/aws-smithy-runtime/src/client/http.rs
@@ -14,7 +14,11 @@ pub mod test_util;
 /// This module is named after the hyper version number since we anticipate
 /// needing to provide equivalent functionality for hyper 1.x in the future.
 #[cfg(feature = "connector-hyper-0-14-x")]
-pub mod hyper_014;
+#[deprecated = "hyper 0.14.x connector is deprecated, please use the connector-hyper-1-x feature instead"]
+pub mod hyper_014 {
+    #[allow(deprecated)]
+    pub use aws_smithy_http_client::hyper_014::*;
+}
 
 /// HTTP body and body-wrapper types
 pub mod body;
